backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas>=2.1.0
numpy>=1.25.0
python-dateutil>=2.8.2
Werkzeug>=3.0.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.2.0 
psycopg2-binary>=2.9.0 